Add 'harvester'
Initial work on a Bash script that will become the Timetasker for
Harvest (using the 'hcl' command line utility).

Here we set up some basic option parsing thanks to this Stack Overflow
answer from 6EQUJ5:
https://stackoverflow.com/questions/22394407/handling-flags-for-a-shell-script-without-getopts/22395652#22395652

In it, a way to parse command-line options is described that works for
both short and long flags.

We'll be using these values to call out to 'hcl'.