harvester: Implement the `--list-modules` flag
When given with a project (project must come first in the argument
list), a list of all the modules (known as "tasks" in Harvest speak)
will be printed to the console, filtered from a list fetched by 'hcl'.

With no project argument, all tasks for all projects will be printed.