harvester: Add `--update-module` flag
This new flag provides an interface to update the current sprint (or
"module" or "task") of a project. Running a command like:

    $ harvester -p proj --update-module

will open `pick` with the list of tasks for that module. The task chosen
from the given list becomes the new sprint associated with the project
alias. This makes it much easier to update sprints.

Move error handling into the more general `exit_with_error` function to
exit with a default error message prefix. The error message is also now
printed to STDERR, which is more standard, and wasn't the case before.